+++ kure_kosen_map/components/map/mymap.vue
@@ -1,0 +1,30 @@
+<template>
+  <div>
+    <div id="mymap" />
+    <layerSwitcher />
+    <popup />
+  </div>
+</template>
+
+<script>
+import initMap from "../../assets/map/initMap";
+
+import layerSwitcher from "./layerSwitcher";
+import popup from "./popup";
+
+export default {
+  name: "MyMap",
+  components: {
+    layerSwitcher,
+    popup
+  },
+  mounted: function() {
+    initMap(this);
+  }
+};
+</script>
+
+<style scoped>
+@import url("../../assets/map/map.css");
+@import url("../../assets/map/map-design.css");
+</style>
