--- kure_kosen_map/components/mymap.vue
@@ -1,21 +1,0 @@
-<template>
-  <div>
-    <div id="mymap" />
-  </div>
-</template>
-
-<script>
-import initMap from "../assets/map/initMap.js";
-
-export default {
-  name: "MyMap",
-  mounted: function() {
-    initMap(this);
-  }
-};
-</script>
-
-<style scoped>
-@import url("../assets/map/map.css");
-@import url("../assets/map/map-design.css");
-</style>
